requirements.txt
beautifulsoup4==4.3.2
Flask==0.10.1
gunicorn==19.2.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-dateutil==2.4.0
requests==2.5.1
six==1.9.0
Werkzeug==0.10.1
